refactor: split mm_api module
